pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.3",
    "filelock>=3.25.2",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openrouter>=0.2.1",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "numpy>=2.4.4",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "requests>=2.33.1",
    "sentence-transformers>=5.3.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.43.0",
]
